requirements.txt
beautifulsoup4
python-dotenv
fastapi
fire
pandas
pydantic
pyodbc
requests
scikit-learn
sqmodel
uvicorn